requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.43.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
